Add support for IPV6
Previous implementation only supported detecting and setting A records
for ipv4. As I'm starting to play a bit with ipv6 and with it's growing
support I find it interesting to enable GandIP users to automatically
update AAAA records for a domain name.

Users can choose to not update the ipv4 or ipv6 with the `--noipv4` and
`--noipv6` options.

I dropped support for setting a custom provider to get the ipv4 and
ipv6. Instead I use the ipify.org [1] API wich provides support for both
ipv4 and ipv6 addresses.

[1] https://www.ipify.org/